PlayerController
Player controller script has been created.
Small refactor.
