+++ TheDepth/Assets/__Scripts/Player/PlayerController.cs
@@ -1,0 +1,31 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerController : MonoBehaviour
+{
+    private Mover mover;
+
+    private void Awake()
+    {
+        mover = GetComponent<Mover>();
+    }
+
+    private void Update()
+    {
+        if (Input.GetMouseButton(0))
+        {
+            MoveToCursor();
+        }
+    }
+
+    private void MoveToCursor()
+    {
+        Ray ray = Camera.main.ScreenPointToRay(Input.mousePosition);
+        bool hasHit = Physics.Raycast(ray, out RaycastHit hit);
+        if (hasHit)
+        {
+            mover.MoveTo(hit.point);
+        }
+    }
+}
